Tests for different kinds of init for both methods
